ready for draft and todo
when a user creates a source, a passthrough pipeline is created for that source, the new passthrough pipeline compiles, and then it's ready for sample viewing.  The user views the sample through the sources list page and the "Eye" icon (inspect).

When the user clicks the refresh button in the inspect view for a source, the pipeline starts, then collects a number (20 right now) of records, displays in the view, and then the pipeline is stopped.