Use SWR for fetching data for job details page
Use the SWR library to provide react hooks for fetching the data shown
on the job details page. All the data fetching logic for that page
(excluding the streaming request for now) has been moved to
data_fetching.ts.

Add an JobError component that's shown if the job can't be fetched,
whether because it doesn't exist or for any other reason.

Also refactor JobDetails.tsx and separate it into several components.

Resolves: #72