Add pipeline endpoints to REST API
Add endpoints for managing pipelines to the REST API. Creating a
pipeline automatically creates a single job for the pipeline.

Note that the REST API depends on the gRPC API so that it can use the
'start_or_preview' function in the POST /v1/pipelines endpoint. When we
are ready to deprecate the gRPC API we can move that logic into
pipelines.rs.