Use REST API for majority of console
Update the console to use the REST API for most data fetching, excluding
connection tables, metrics, checkpoint details, and subscribing to job
output. Mildly redesigns the pipelines page to make the distinction
between pipelines and jobs more clear.

Note that this also changes the remaining GRPC endpoints to operate on
the public ids that the REST API exposes. The GRPC endpoints internally
look up the internal id and then operate on that.

Also remove global error handling and explicitly handle the API being
unavailable by polling the /v1/ping endpoint. The page displays a
message if this request fails.