Switch all connection endpoints to REST API
Deprecate all the GRPC methods for managing connections (now called
'connection profiles') and connection tables and move them to the REST
API. Switch the console to use the new endpoints.

Also fix some styling issues in the console, including limiting the
error message on the create pipeline page to 2 lines.