Fix how we set subtask metadata has_state
The subtask metadata has_state was being set based on whether new bytes
were written during the epoch being checkpointed, which is insufficient
because it's possible for no new bytes to be written. This caused an
error later during compaction because when we write operator checkpoint
metadata we exclude subtasks that don't have state. The result was files
being excluded from the operator metadata.