Support deletes in the parquet backend
Modify the parquet schema to support writing deletions. Currently only
the KeyedState table performs any deletions. Reading the parquet file
when restoring from a checkpoint accounts for deletes.

This also adds a new mechanism for generational compaction. After
completing a checkpoint, the controller will look at the backend and for
each operator and table, if there are at least MIN_FILES_TO_COMPACT
(currently set to 4) of the same generation, it will read those files,
account for deletes, and write a set of newly partitioned, compacted
files. The controller then calls an rpc on the workers to trigger them
to load that new data.

Compaction defaults to disabled, and can be enable with the environment
variable COMPACTION_ENABLED.