Allow compaction on previously compacted epochs
If the controller is restarted immediately after a compaction it will
try to compact the same epochs as its last compaction because the latest
checkpoint specifies the min epoch, and that hasn't yet been updated.
This allows compaction to continue even if the files it tries to delete
don't exist.