Run cargo check on udfs
Add an API endpoint that accepts the contents of the UDFs editor and
sends it to the controller, and in turn to the compiler service, which
runs 'cargo check'. Pass any errors back to the API and show them in the
console. Also show the query validation errors in the Errors tab for
consistency.

Note that the UDFs are put into a new 'udfs' crate in the build
directory for the purpose of cargo check, and written again the
'pipeline' crate for the actual pipeline. A future commit will change
the pipeline to use the udfs crate.