Improve UDF state handling
If a UDF validation request is in progress while the user continues
editing the definition input, the request should be aborted.

Also keep the state of the UDF editor separate from local storage to
make the data flow in only one direction.

Also changes the UDF validation fetcher to run on empty definitions.