Make UDFs public when added to schema provider
Restore the previous logic of modifying the UDF syntax tree at the time
we add it to the schema provider.

Also use the schema provider in the check UDF flow. This removes some
redundant logic from the controller and improves the error messages that
now appear as UDF validation errors with their respective UDF rather
than as just as query errors.