Support environment variable substitution
In the Kafka, Polling HTTP, WebSocket, and SSE sources, support
environment variables in the passwords, secrets, and headers fields.

This is done by setting the format of the field in the JSON schema to
'var-str', which tells typify to replace the type with the VarStr
struct, which has a function for substituting environment variables.
The VarStr struct serializes/deserializes from a JSON string.