Fix issues with enums in json form
Formik was not properly setting the values when switching between JSON
shema 'oneof' variants. This adds a call to reset the field when
switching variants.

This change also removes the option of leaving optional enums as null,
and instead defaults it to the 'default_value' (for string enums if
defined), or the first oneof variant.