Generate watermarks and set event time using expressions
Modify the periodic watermark generator to accept a physical expressions
that it uses to calculate the watermark. The expression can be set via a
virtual field in the create table statement, or it defaults to a max
lateness expression. The operator is inserted via a user-defined logical
node.

Add an expression to the projection in the SourceRewriter to set the
event time if present.