Load dynamically linked UDF libraries at runtime
Build a wrapper library for each UDF that gets loaded by the worker
at runtime. The library code is generated at pipeline creation time, and
the worker calls the UDF via the foreign function interface.

This change also brings back the compiler service, converting it to a
library and launching the service with arroyo-bin.